+++ packages/fs-detectors/test/unit.builds-and-routes-detector.test.ts
@@ -156,7 +156,7 @@ describe('Test `detectBuilders`', () => {
     );
   });
 
-  it('should use services builders when experimentalServicesV2 is configured without the services framework', async () => {
+  it('should use services builders when services is configured without the services framework', async () => {
     const workPath = join(
       __dirname,
       'fixtures',
@@ -165,7 +165,7 @@ describe('Test `detectBuilders`', () => {
     );
     const { builders, services, errors, useImplicitEnvInjection } =
       await detectBuilders([], undefined, {
-        experimentalServicesV2: {
+        services: {
           web: {
             root: '.',
             runtime: 'python',